consolidate import_google()
